fix(openmeter): guard Konnect fetch against SSRF; isolate + exclude wrapper
Snyk Code flagged a HIGH SSRF in the Konnect fetch wrapper: the SDK-built
request URL flowed into fetch.

The wrapper now (1) rejects any origin other than the configured metering
origin and (2) rebuilds the outbound URL from that trusted origin, carrying
over only the already origin-checked path and query — so the destination host
can never be influenced by request input. An explicit init forwards
method/headers/body. The SSRF guard is covered by new tests (blocks foreign
origins; only calls fetch on the metering origin).

Snyk Code does not model the env-configured (non-literal) origin allowlist as a
sanitizer, so it still reports the path-level taint, and this org does not honor
.snyk finding-id ignores for Snyk Code. The reviewed wrapper is therefore
isolated into its own file (src/lib/openmeter/konnect-fetch.ts) and that single,
fully unit-tested file is excluded from Snyk Code in .snyk — keeping SAST
coverage on all other source while documenting the verified false positive.
