+++ src/lib/openmeter/client.ts
@@ -1,115 +1,9 @@
 import { OpenMeter } from "@openmeter/sdk";
 import { getHostedOpenMeterUrl, isOpenMeterEnabled } from "./constants";
-import { resolveKonnectMeterId } from "./konnect-catalog";
-import {
-  buildKonnectMeterQueryBody,
-  isKonnectMeterQueryGet,
-  isKonnectMeterQueryPath,
-  normalizeKonnectMeterQueryResponse,
-  normalizeKonnectResponseBody,
-  rewriteKonnectRequestBody,
-  rewriteKonnectRequestUrl,
-} from "./konnect-routes";
-import { resolveHostedOpenMeterBaseUrl, shouldUseKonnectRoutes, assertKonnectFetchOrigin } from "./route-mode";
+import { createKonnectFetch } from "./konnect-fetch";
+import { resolveHostedOpenMeterBaseUrl, shouldUseKonnectRoutes } from "./route-mode";
 
 let hostedClient: OpenMeter | null = null;
-
-const KONNECT_METER_QUERY_PATH_RE = /\/meters\/([^/]+)\/query$/;
-
-async function buildKonnectRequest(request: Request): Promise<Request> {
-  const sourceUrl = new URL(request.url);
-  const rewritten = rewriteKonnectRequestUrl(sourceUrl, request.method);
-
-  if (isKonnectMeterQueryGet(sourceUrl.pathname, request.method)) {
-    const body = buildKonnectMeterQueryBody(sourceUrl.searchParams);
-    const headers = new Headers(request.headers);
-    headers.set("content-type", "application/json");
-    rewritten.search = "";
-
-    const meterMatch = KONNECT_METER_QUERY_PATH_RE.exec(rewritten.pathname);
-    if (meterMatch) {
-      const meterId = await resolveKonnectMeterId(meterMatch[1]);
-      rewritten.pathname = rewritten.pathname.replace(
-        `/meters/${meterMatch[1]}/query`,
-        `/meters/${meterId}/query`,
-      );
-    }
-
-    return new Request(rewritten.toString(), {
-      method: "POST",
-      headers,
-      body: JSON.stringify(body),
-      redirect: request.redirect,
-      signal: request.signal,
-      credentials: request.credentials,
-      integrity: request.integrity,
-      keepalive: request.keepalive,
-      mode: request.mode,
-      referrer: request.referrer,
-      referrerPolicy: request.referrerPolicy,
-    });
-  }
-
-  const contentType = request.headers.get("content-type") ?? "";
-  if (
-    request.method !== "GET" &&
-    request.method !== "HEAD" &&
-    contentType.includes("application/json")
-  ) {
-    const reqClone = request.clone();
-    try {
-      const json = await request.json();
-      const body = rewriteKonnectRequestBody(rewritten.pathname, request.method, json);
-      return new Request(rewritten.toString(), {
-        method: request.method,
-        headers: request.headers,
-        body: JSON.stringify(body),
-        redirect: request.redirect,
-        signal: request.signal,
-        credentials: request.credentials,
-        integrity: request.integrity,
-        keepalive: request.keepalive,
-        mode: request.mode,
-        referrer: request.referrer,
-        referrerPolicy: request.referrerPolicy,
-      });
-    } catch {
-      return new Request(rewritten.toString(), reqClone);
-    }
-  }
-
-  return new Request(rewritten.toString(), request);
-}
-
-function createKonnectFetch(allowedBaseUrl: string): typeof fetch {
-  return async (input, init) => {
-    const request = input instanceof Request ? input : new Request(input, init);
-    const konnectRequest = await buildKonnectRequest(request);
-    assertKonnectFetchOrigin(konnectRequest.url, allowedBaseUrl);
-    const response = await fetch(konnectRequest);
-    const contentType = response.headers.get("content-type") ?? "";
-    if (!contentType.includes("application/json")) {
-      return response;
-    }
-    const body = await response.json();
-    let normalized = normalizeKonnectResponseBody(body);
-    if (isKonnectMeterQueryPath(new URL(konnectRequest.url).pathname)) {
-      normalized = normalizeKonnectMeterQueryResponse(normalized);
-    }
-    if (normalized === body) {
-      return new Response(JSON.stringify(body), {
-        status: response.status,
-        statusText: response.statusText,
-        headers: response.headers,
-      });
-    }
-    return new Response(JSON.stringify(normalized), {
-      status: response.status,
-      statusText: response.statusText,
-      headers: response.headers,
-    });
-  };
-}
 
 export function createOpenMeterClient(input: {
   baseUrl: string;
